moving memoization into a class so it's not global
