+++ experimental/reporting-ui/src/main/react/reporting-ui/library/reportingClient/models/GetReportResponse.ts
